// shared/types.ts
export type TreeTypes = {
  pineSeed: number;
  pineShoot: number;
  pineWood: number;
  pineNeedle: number;
  rawWood: number;
  charcoal: number;
  leafDust: number;
  wetSand: number;
  fire: number;
  flame: number;
};

export type TreeFields = {
  rootX: number;
  rootY: number;
  progress: number;
  phase: number;
  seedCount: number;
  halfWidth: number;
  targetHeight: number;
};